fix sums were not correct
That happened because before this change the sums where calculated with the bytes of the content of the file, but when you take the sum of a regular file, the sum is also calculated with the metadata, not only the content. (at least thats my theory). this has been fixed with this patch.