add safety checker
This is a prototype of a safety checker, which should help to prevent shooting yourself in the foot. If you are sure of what you are doing you can opt out of the checker by using the `--bypass-all-checks` flag, but it is more recommended using the new `unsafe` token, which will skip individual functions. It can be used like this: `copy(@name, @name.txt) unsafe print(@sum=sha256)`. I am not sure if maintaining both options is too much overkill, though.